+++ cpp/date_util_module/UtilDate.cpp
@@ -12,7 +12,7 @@
 #include <chrono>
 #include <ctime>
 #include <mgp.hpp>
-#include "mchputil/DateHelper.hpp"
+#include "util/DateHelper.hpp"
 #include <string>
