--- cpp/mchp_util_module/CMakeLists.txt
@@ -1,6 +1,0 @@
-set(mchp_util_date_src
-    MchpUtilDate.cpp
-    mchputil/DateHelper.cpp)
-
-
-add_query_module(MchpUtilDate 1 "${mchp_util_date_src}")
